backend/requirements.txt
fastapi
motor
python-dotenv
python-multipart
